[table] don't use parse multi, it cause trouble here.
the reason is we don't save config option if it's same as default value,
but, the system config might not use default so parse multi could end
up with read a system value instead of user choosed value.

the initial intention is to simplify how to disable a table, but since
now a much more convinient im list ui is there, we don't need to do that
anymore.